requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
metadata-parser==0.10.4
mistune==0.8.4
Pillow==8.0.1
python-frontmatter==0.5.0
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.2
